🌟 ansible: deepen project model + fix dict integer serialization
Builds on the project-wide analysis with correctness fixes and broader coverage.

Correctness:
- Parameterized roles (`roles: [{role: x, when: ..., tags: ..., vars: ...}]`) no
  longer fail to parse. Play roles are exposed as ansible.play.roleApplication
  carrying the application directives plus the resolved role.
- group_vars/host_vars are now honored beside the inventory source
  (e.g. inventory/group_vars/), not only at the project root.
- Vault: the header vault-id is captured, and variables encrypted inline with a
  `!vault` tag are detected (ansible.vault.variables) in addition to fully
  encrypted files.

New surface:
- ansible.task.module / moduleArgs — the module a task invokes and its args,
  so audits select by module without parsing the action dict.
- ansible.project.plugins — custom modules and plugins (library/, *_plugins/,
  plugins/<type>/), a supply-chain surface.
- ansible.project.collections — collections vendored under
  collections/ansible_collections/, and ansible.project.manifest for a project
  that is itself a collection (galaxy.yml).
- ansible.role.argumentSpecs — the role's declared input contract.
- Play-level varsFiles, varsPrompt, environment, collections.
- ansible.project.lintConfig and moleculeScenarios as quality signals.

Fix: dict fields sourced from YAML stored Go `int`, which the llx dict-to-
primitive conversion rejects ("unsupported child type: int") — any integer in a
vars/defaults/action/moduleArgs dict broke rendering. Values are now normalized
to int64/float64 before storage, which also fixes the pre-existing play.vars and
task.action cases.

+++ providers/ansible/resources/ansible_test.go
@@ -237,7 +237,7 @@ func TestSerialIsExposed(t *testing.T) {
 	plays, err := a.plays()
 	require.NoError(t, err)
 	require.Len(t, plays, 1)
-	assert.Equal(t, 3, plays[0].(*mqlAnsiblePlay).Serial.Data)
+	assert.Equal(t, int64(3), plays[0].(*mqlAnsiblePlay).Serial.Data)
 }
 
 // Task-level security attributes (become family, no_log, ignore_errors,
